+++ apps/web/src/middleware.ts
@@ -14,6 +14,7 @@ export function middleware(request: NextRequest) {
 
   if (
     pathname.startsWith('/_next') ||
+    pathname.startsWith('/api') ||
     pathname.startsWith('/favicon') ||
     pathname.includes('.')
   ) {